Do something :) Add fullWaterOrGaz class and some experimental
