+++ src/main/java/project/factory/Stocktaking.java
@@ -1,0 +1,4 @@
+package main.java.project.factory;
+
+public class Stocktaking {
+}
