--- src/main/java/project/material/Sand.java
@@ -1,10 +1,0 @@
-package main.java.project.material;
-
-public class Sand implements Transformable{
-
-    @Override
-    public void mix() {
-
-    }
-    
-}
