+++ src/main/java/project/stuff/Sand.java
@@ -1,0 +1,10 @@
+package main.java.project.stuff;
+
+public class Sand implements Transformable {
+
+    @Override
+    public void mix() {
+
+    }
+    
+}
